move error and 400 page to a commponents
