+++ src/components/error/errorRespond.tsx
@@ -1,18 +1,10 @@
 'use client'
-
 import Typography from '@/components/ui/Typography'
 import { Button } from '@/components/ui/button'
-import absurd from '../../public/absurd 2.png'
+import absurd from '../../../public/absurd 2.png'
 import Image from 'next/image'
 import { useRouter } from 'next/navigation'
-
-export default function Error({
-    error,
-    reset,
-}: {
-    error: Error & { digest?: string }
-    reset: () => void
-}) {
+export default function ErrorRespond({ reset }: { reset: () => void }) {
     const router = useRouter()
 
     return (
@@ -21,11 +13,15 @@ export default function Error({
             <div className="flex w-fit flex-col justify-self-end">
                 <Typography level="H1">400</Typography>
                 <Typography level="P">
-                    Ser ut som att du försöker nå en sida som inte finns.
+                    Det blev något fel på våran sida. Försök igen senare eller
+                    prova ladda om sidan.
                 </Typography>
                 <div className="mt-6 flex gap-11">
+                    <Button variant="outline" onClick={() => reset()}>
+                        Ladda om sidan
+                    </Button>
                     <Button onClick={() => router.push('/')}>
-                        Gå till start sidan
+                        Gå till startsidan
                     </Button>
                 </div>
             </div>
