+++ src/components/error/index.tsx
@@ -1,0 +1,4 @@
+import ErrorRespond from './errorRespond'
+import NotfoundRespond from './notfoundRespond'
+
+export { ErrorRespond, NotfoundRespond }
